# docs/swagger.yaml
definitions:
  model.Account:
    properties:
      createdAt:
        type: string
      description:
        maxLength: 50
        minLength: 2
        type: string
      id:
        maxLength: 32
        type: string
      projectID:
        maxLength: 32
        type: string
      updatedAt:
        type: string
    required:
    - description
    - id
    - projectID
    type: object
  model.Accounts:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Account'
        type: array
      pager:
        $ref: '#/definitions/pkg.Pager'
    type: object
  model.Event:
    properties:
      createdAt:
        type: string
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/model.Field'
        type: array
      id:
        maxLength: 32
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      onTrail:
        description: 这个字段暂时用不上
        type: boolean
      recordIDs:
        description: 不要在数据库中保存这个字段
        items:
          type: string
        type: array
      records:
        items:
          $ref: '#/definitions/model.Record'
        type: array
      results:
        items:
          $ref: '#/definitions/model.EventResult'
        type: array
      storyID:
        maxLength: 32
        type: string
      updatedAt:
        type: string
    required:
    - id
    - name
    - storyID
    type: object
  model.EventLog:
    properties:
      content:
        type: string
      createdAt:
        type: string
      eventID:
        maxLength: 32
        type: string
      fieldLogs:
        items:
          $ref: '#/definitions/model.FieldLog'
        type: array
      id:
        description: jet返回时每个eventLog都带id
        maxLength: 32
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      platform:
        type: string
      raw:
        type: string
      recordID:
        maxLength: 32
        type: string
      tested:
        description: 被测试用
        type: boolean
      updatedAt:
        type: string
      used:
        description: 被使用过（被前端clear log）
        type: boolean
      userId:
        description: 既然有这个信息，不妨存一下
        type: string
    required:
    - eventID
    - id
    - name
    - recordID
    type: object
  model.EventResult:
    properties:
      android:
        $ref: '#/definitions/model.TestResult'
      createdAt:
        type: string
      eventID:
        type: string
      id:
        maxLength: 32
        type: string
      ios:
        $ref: '#/definitions/model.TestResult'
      other:
        $ref: '#/definitions/model.TestResult'
      recordID:
        type: string
      updatedAt:
        type: string
    required:
    - id
    type: object
  model.Field:
    properties:
      createdAt:
        type: string
      description:
        type: string
      eventID:
        maxLength: 32
        type: string
      id:
        maxLength: 32
        type: string
      key:
        type: string
      results:
        items:
          $ref: '#/definitions/model.FieldResult'
        type: array
      type:
        type: string
      typeID:
        type: string
      updatedAt:
        type: string
      value:
        description: 插入时会把数组用"|"隔开变成字符串，读取的时候也要转化一下
        type: string
    required:
    - eventID
    - id
    type: object
  model.FieldLog:
    properties:
      createdAt:
        type: string
      eventLogID:
        maxLength: 32
        type: string
      fieldID:
        maxLength: 32
        type: string
      id:
        maxLength: 32
        type: string
      key:
        type: string
      platform:
        type: string
      tested:
        description: 被测试用 //这个字段暂时用不上
        type: boolean
      updatedAt:
        type: string
      used:
        description: 被使用过（被前端clear log） //这个字段暂时用不上
        type: boolean
      value:
        type: string
    required:
    - eventLogID
    - fieldID
    - id
    type: object
  model.FieldResult:
    properties:
      android:
        $ref: '#/definitions/model.TestResult'
      createdAt:
        type: string
      fieldID:
        type: string
      id:
        maxLength: 32
        type: string
      ios:
        $ref: '#/definitions/model.TestResult'
      other:
        $ref: '#/definitions/model.TestResult'
      recordID:
        type: string
      updatedAt:
        type: string
    required:
    - id
    type: object
  model.Project:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        maxLength: 500
        minLength: 2
        type: string
      updatedAt:
        type: string
    required:
    - name
    type: object
  model.Projects:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      pager:
        $ref: '#/definitions/pkg.Pager'
    type: object
  model.Record:
    properties:
      createdAt:
        type: string
      eventLogs:
        items:
          $ref: '#/definitions/model.EventLog'
        type: array
      events:
        description: Events    string     `json:"events"` // 存event的id数组
        items:
          $ref: '#/definitions/model.Event'
        type: array
      filter:
        type: string
      id:
        maxLength: 32
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      projectID:
        maxLength: 32
        type: string
      status:
        $ref: '#/definitions/model.status'
      updatedAt:
        type: string
    required:
    - id
    - name
    - projectID
    type: object
  model.Stories:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Story'
        type: array
      pager:
        $ref: '#/definitions/pkg.Pager'
    type: object
  model.Story:
    properties:
      createdAt:
        type: string
      description:
        type: string
      events:
        items:
          $ref: '#/definitions/model.Event'
        type: array
      id:
        maxLength: 32
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      projectID:
        maxLength: 32
        type: string
      updatedAt:
        type: string
    required:
    - id
    - name
    - projectID
    type: object
  model.SwaggerEventLogs:
    properties:
      data:
        items:
          $ref: '#/definitions/model.EventLog'
        type: array
      pager:
        $ref: '#/definitions/pkg.Pager'
    type: object
  model.SwaggerEvents:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Event'
        type: array
      pager:
        $ref: '#/definitions/pkg.Pager'
    type: object
  model.TestResult:
    enum:
    - SUCCESS
    - FAIL
    - UNCERTAIN
    - UNTESTED
    type: string
    x-enum-varnames:
    - SUCCESS
    - FAIL
    - UNCERTAIN
    - UNTESTED
  model.status:
    enum:
    - "ON"
    - "OFF"
    type: string
    x-enum-varnames:
    - "ON"
    - "OFF"
  pkg.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
  pkg.Pager:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      totalRow:
        type: integer
    type: object
info:
  contact: {}
  title: TrackMaster
  version: "2.0"
paths:
  /api/v2/accounts:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: page size
        in: query
        name: pageSize
        type: string
      - description: description
        in: query
        name: description
        type: string
      - description: project id
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Accounts'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - account
    post:
      parameters:
      - description: project ID
        in: body
        name: projectID
        required: true
        schema:
          type: string
      - description: page size
        in: body
        name: id
        required: true
        schema:
          type: string
      - description: description
        in: body
        name: description
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Account'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - account
  /api/v2/accounts/{id}:
    delete:
      parameters:
      - description: account id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - account
  /api/v2/projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Projects'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - project
  /api/v2/projects/sync:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - project
  /api/v2/realTime/clearLog:
    post:
      parameters:
      - description: record ID
        in: query
        name: record
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/getLog:
    get:
      parameters:
      - description: record ID
        in: query
        name: record
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.SwaggerEventLogs'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/getResult:
    get:
      parameters:
      - description: record ID
        in: query
        name: record
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.SwaggerEvents'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/start:
    post:
      parameters:
      - description: project ID
        in: body
        name: project
        required: true
        schema:
          type: string
      - description: account IDs
        in: body
        name: accounts
        required: true
        schema:
          items:
            type: string
          type: array
      - description: event IDs
        in: body
        name: events
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Record'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/stop:
    post:
      parameters:
      - description: record ID
        in: query
        name: record
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Record'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/update:
    post:
      parameters:
      - description: record ID
        in: query
        name: record
        required: true
        type: string
      - description: account IDs
        in: body
        name: accounts
        schema:
          items:
            type: string
          type: array
      - description: event IDs
        in: body
        name: events
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Record'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/realTime/updateResult:
    post:
      parameters:
      - description: record ID
        in: body
        name: record
        required: true
        schema:
          type: string
      - description: field IDs
        in: body
        name: fields
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Field'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - realTime
  /api/v2/stories:
    get:
      parameters:
      - description: page, default 10
        in: query
        name: page
        type: string
      - description: page size, default 10
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Stories'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - story
  /api/v2/stories/{id}:
    get:
      parameters:
      - description: story id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Story'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - story
  /api/v2/stories/sync:
    post:
      parameters:
      - description: project id
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Story'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/pkg.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/pkg.Error'
      tags:
      - story
swagger: "2.0"